BAH-4552 | Refactor. Rename ImmunizationHistory Component As ImmunizationForm Component
