+++ apps/clinical/src/components/forms/immunization/__tests__/SelectedImmunizationItem.test.tsx
@@ -16,7 +16,7 @@ import {
   mockRoutesValueSet,
   mockSitesValueSet,
   mockStore,
-} from './__mocks__/immunizationHistoryMocks';
+} from './__mocks__/immunizationMocks';
 
 jest.mock('../stores');
